BoardDrawer 切り出し
