+++ src/components/index.ts
@@ -9,6 +9,7 @@ export { SnackbarProvider } from './common/SnackbarProvider'
 export { Board } from './views/Board'
 export { Home } from './views/Home'
 export { ArchivedBoardModal } from './board/ArchivedBoardModal'
+export { BoardDrawer } from './board/BoardDrawer'
 export { BoardListSearchForm } from './board/BoardListSearchForm'
 export { BoardTitle } from './board/BoardTitle'
 export { CreateBoardModal } from './board/CreateBoardModal'
