+++ src/components/views/Board.tsx
@@ -6,10 +6,8 @@ import * as I from '@/scripts/model/interface'
 import { LoadingSpinner, CardList } from '@/components'
 import { useSnackbarContext } from '@/scripts/hooks'
 import { css } from '@emotion/core'
-import { BoardTitle } from '@/components'
+import { BoardTitle, BoardDrawer } from '@/components'
 import { Button } from '@material-ui/core'
-import { Drawer, makeStyles } from '@material-ui/core'
-import { MoreHoriz } from '@material-ui/icons'
 import { Theme } from '@material-ui/core'
 import { fetchList, createList } from '@/scripts/redux/state/list/actions'
 
@@ -22,10 +20,7 @@ export const Board: React.FC = () => {
   const listState = useSelector((state: I.ReduxState) => state.list)
   const dispatch = useDispatch()
   const { showSnackbar } = useSnackbarContext()
-  const [open, setOpen] = React.useState(false)
-  const muiStyle = useStyles()
-  const history = useHistory()
-  const { boardId } = useParams<{ boardId: string }>()
+  const { boardId } = useParams<I.UrlParams>()
 
   React.useEffect(() => {
     ;(async () => {
@@ -55,22 +50,6 @@ export const Board: React.FC = () => {
       })()
     }
   }, [dispatch, userState, showSnackbar])
-
-  const toggleDrawer = () => {
-    setOpen(!open)
-  }
-
-  const onClickArchive = async () => {
-    if (!boardId) return
-    if (!window.confirm('ボードをアーカイブしてもよろしいですか？')) return
-    try {
-      await dispatch(archiveBoard({ id: boardId }))
-      setOpen(false)
-      history.push('/boards')
-    } catch ({ message }) {
-      showSnackbar({ message, type: 'error' })
-    }
-  }
 
   const onClick = () => {
     if (boardId) dispatch(createList({ title: 'new card', boardId }))
@@ -107,62 +86,16 @@ export const Board: React.FC = () => {
           )}
         </>
       )}
-      <Drawer
-        anchor="right"
-        open={open}
-        onClose={toggleDrawer}
-        className={muiStyle['root']}
-        variant="persistent"
-      >
-        <div css={styles['drawer-button']}>
-          <Button
-            onClick={toggleDrawer}
-            variant="contained"
-            startIcon={<MoreHoriz />}
-          >
-            ボードメニューを表示
-          </Button>
-        </div>
-        <div css={styles['drawer-content']}>
-          <Button
-            onClick={onClickArchive}
-            fullWidth
-            className={muiStyle['archiveButton']}
-          >
-            このボードをアーカイブ
-          </Button>
-        </div>
-      </Drawer>
+      <BoardDrawer />
     </div>
   )
 }
-
-const useStyles = makeStyles(() => ({
-  root: {
-    '& .MuiDrawer-paper': {
-      width: 300,
-      overflow: 'visible',
-      top: 64,
-      padding: 8,
-      boxShadow: '2px 0px 7px 0px #c1c1c1'
-    }
-  },
-  archiveButton: {
-    textDecoration: 'underline'
-  }
-}))
 
 const styles = {
   root: css`
     padding: 10px;
     position: relative;
   `,
-  'drawer-button': css`
-    position: absolute;
-    left: -76%;
-    visibility: visible;
-  `,
-  'drawer-content': css``,
   'board-header': (theme: Theme) => css`
     margin-bottom: ${theme.spacing(1)}px;
   `,
